Rename everything to snakey cases
