+++ src/app/(app)/my-decks/[lang]/browse.jsx
@@ -3,8 +3,8 @@
 import Link from 'next/link'
 import { useState } from 'react'
 import Select from 'react-select'
-import Loading from 'app/loading'
-import ErrorList from 'components/ErrorList'
+import Loading from 'components/loading'
+import ErrorList from 'components/error-list'
 import { useLanguageDetails } from 'app/data/hooks'
 import BigPhrase from 'components/BigPhrase'
