+++ src/app/(app)/my-decks/[lang]/new-card/page.jsx
@@ -1,6 +1,6 @@
 import languages from 'lib/languages'
 import AddCardPhraseForm from './form'
-import Navbar from 'app/(app)/Navbar'
+import Navbar from 'app/(app)/navbar'
 
 export default function Page({ params: { lang } }) {
   return (
