+++ src/app/(app)/my-decks/new/page.jsx
@@ -1,5 +1,5 @@
 import Form from './form'
-import Navbar from 'app/(app)/Navbar'
+import Navbar from 'app/(app)/navbar'
 
 export default function Page() {
   return (
