+++ src/app/(app)/profile/page.jsx
@@ -1,5 +1,5 @@
 import Link from 'next/link'
-import Navbar from 'app/(app)/Navbar'
+import Navbar from 'app/(app)/navbar'
 import UpdateProfileForm from './update-profile-form'
 import UserAuthCard from './user-auth-card'
