+++ src/app/(app)/profile/update-profile-form.tsx
@@ -11,9 +11,9 @@ import { useQueryClient } from '@tanstack/react-query'
 import { toast } from 'react-hot-toast'
 import supabase from 'lib/supabase-client'
 import languages from 'lib/languages'
-import Error from 'components/Error'
+import Error from 'components/error'
 import SelectMultipleLanguagesInput from 'components/select-multiple-languages'
-import Loading from 'app/loading'
+import Loading from 'components/loading'
 import { useProfile } from 'app/data/hooks'
 import AvatarEditor from './avatar-edit'
