+++ src/app/(auth)/forgot-password/form.jsx
@@ -2,7 +2,7 @@
 
 import { useState } from 'react'
 import supabase from 'lib/supabase-client'
-import ErrorList from 'components/ErrorList'
+import ErrorList from 'components/error-list'
 import { useMutation } from '@tanstack/react-query'
 import { BASE_URL } from 'lib/helpers'
