+++ src/app/(auth)/set-new-password/form.jsx
@@ -3,7 +3,7 @@
 import { useMutation } from '@tanstack/react-query'
 import Link from 'next/link'
 import supabase from 'lib/supabase-client'
-import ErrorList from 'components/ErrorList'
+import ErrorList from 'components/error-list'
 import { useAuth } from 'lib/auth-context'
 import { toast } from 'react-hot-toast'
