+++ src/app/(auth)/signup/form.jsx
@@ -4,7 +4,7 @@ import { useMutation } from '@tanstack/react-query'
 import Link from 'next/link'
 import { useRouter } from 'next/navigation'
 import supabase from 'lib/supabase-client'
-import ErrorList from 'components/ErrorList'
+import ErrorList from 'components/error-list'
 import { BASE_URL } from 'lib/helpers'
 
 export default function SignupForm() {
