+++ src/app/getting-started/client.jsx
@@ -4,14 +4,14 @@ import { useState } from 'react'
 import { useMutation } from '@tanstack/react-query'
 import { useAuth } from 'lib/auth-context'
 import supabase from 'lib/supabase-client'
-import Navbar from 'app/(app)/Navbar'
-import ErrorList from 'components/ErrorList'
+import Navbar from 'app/(app)/navbar'
+import ErrorList from 'components/error-list'
 import Link from 'next/link'
 import { useProfile } from 'app/data/hooks'
 import { prependAndDedupe } from 'lib/helpers'
 import { useQueryClient } from '@tanstack/react-query'
 import languages from 'lib/languages'
-import SuccessCheckmark from 'components/SvgComponents'
+import SuccessCheckmark from 'components/svg-components'
 
 export const dynamic = 'force-dynamic'
