+++ src/components/sidebar.jsx
@@ -2,12 +2,12 @@
 
 import { useState, useEffect } from 'react'
 import Link from 'next/link'
-import { Garlic } from 'components/Garlic'
+import { Garlic } from 'components/garlic'
 import languages from 'lib/languages'
 import { usePathname, useRouter } from 'next/navigation'
 import { useProfile } from 'app/data/hooks'
-import Loading from 'app/loading'
-import ErrorList from './ErrorList'
+import Loading from 'components/loading'
+import ErrorList from './error-list'
 import supabase from 'lib/supabase-client'
 import { toast } from 'react-hot-toast'
 import { useAuth } from 'lib/auth-context'
